requirements.txt
asgiref==3.10.0
bytez==3.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-allauth==65.13.1
django-pwa==2.0.1
django-widget-tweaks==1.5.0
idna==3.11
pycparser==2.23
PyJWT==2.10.1
PyPDF2==3.0.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
 